Refactor proxy_conn into a module for sharing
